requirements.txt
colorama==0.4.4
dhooks==1.1.4
discord.py==1.7.3
discord_webhook==0.14.0
pyfiglet==0.8.post1
termcolor==1.1.0
